[feat] 댓글 삭제 권한 설정 (alert 미해결)

+++ src/main/java/com/gdsc/soyu/service/comments/CommentService.java
@@ -1,5 +1,6 @@
 package com.gdsc.soyu.service.comments;
 
+import com.gdsc.soyu.config.auth.dto.SessionUser;
 import com.gdsc.soyu.domain.comment.Comment;
 import com.gdsc.soyu.domain.comment.CommentRepository;
 import com.gdsc.soyu.domain.posts.Posts;
@@ -13,7 +14,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.transaction.Transactional;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.Writer;
+import java.net.URLEncoder;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -41,11 +47,23 @@ public class CommentService {
 
     /* DELETE */
     @Transactional
-    public void delete(Long id) {
-        Comment comment = commentRepository.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("해당 댓글이 존재하지 않습니다. id=" + id));
+    public void delete(HttpServletResponse response, SessionUser loginUser, Long commentId) throws IOException {
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() ->
+                new IllegalArgumentException("해당 댓글이 존재하지 않습니다. id=" + commentId));
 
-        commentRepository.delete(comment);
+        response.setContentType("application/json; charset=utf-8");
+        PrintWriter out = response.getWriter();
+
+        if (comment.getUser().getEmail().equals(loginUser.getEmail())){
+            commentRepository.delete(comment);
+
+            out.println("<script>alert('댓글이 삭제되었습니다.'); </script>");
+            out.flush();
+        }
+        else {
+            out.println("<script>alert('댓글 삭제 권한이 없습니다.'); </script>");
+            out.flush();
+        }
     }
 
 }
